Address review comments
- use Collections.emptyList() instead of new ArrayList()
- handle the temp. key cleanup in a `finally` block
- move the new runner calls to a separate class
