+++ java/code/src/com/suse/manager/webui/services/impl/runner/MgrUtilRunner.java
@@ -159,52 +159,4 @@ public class MgrUtilRunner {
                         new TypeToken<Map<Boolean, String>>() { });
         return call;
     }
-
-    /**
-     * {@link RunnerCall} for writing into a text file.
-     * The result of the call is a text description with how many lines were written.
-     *
-     * @param absolutePath the file path
-     * @param contents the file text contents
-     * @return the runner call
-     */
-    public static RunnerCall<String> writeTextFile(String absolutePath, String contents) {
-        return new RunnerCall("salt.cmd", Optional.of(Map.of(
-                "fun", "file.write",
-                "path", absolutePath,
-                "args", contents)),
-                new TypeToken<String>() { });
-    }
-
-    /**
-     * {@link RunnerCall} for setting mode of given file.
-     * The result of the call is a text description of the file after the operation.
-     *
-     * @param absolutePath the absolute path of the file
-     * @param modeString the desired mode
-     * @return the runner call
-     */
-    public static RunnerCall<String> setFileMode(String absolutePath, String modeString) {
-        return new RunnerCall("salt.cmd", Optional.of(Map.of(
-                "fun", "file.set_mode",
-                "path", absolutePath,
-                "mode", modeString)),
-                new TypeToken<String>() { }
-        );
-    }
-
-    /**
-     * {@link RunnerCall} for removing a file
-     * The result of the call is the boolean, which is true, if path was found and the file deletion succeeded.
-     *
-     * @param absolutePath the absolute path of the file
-     * @return the runner call
-     */
-    public static RunnerCall<Boolean> removeFile(String absolutePath) {
-        return new RunnerCall("salt.cmd", Optional.of(Map.of(
-                "fun", "file.remove",
-                "path", absolutePath)),
-                new TypeToken<Boolean>() { });
-    }
-
 }
